Refactor code to "view" package Avoid initializing Hashtable and use HashMap with getOrDefault. Introduce checkstyle (will not fail build if there are violations)
