+++ src/main/java/com/github/h2002044/lc2/view/DataFlow.java
@@ -1,4 +1,6 @@
-package com.github.h2002044.lc2;
+package com.github.h2002044.lc2.view;
+
+import com.github.h2002044.lc2.Processor;
 
 import java.awt.*;
