+++ src/main/java/com/github/h2002044/lc2/view/ExampleFileFilter.java
@@ -1,4 +1,4 @@
-package com.github.h2002044.lc2;
+package com.github.h2002044.lc2.view;
 
 import javax.swing.filechooser.FileFilter;
 import java.io.File;
