+++ src/main/java/com/github/h2002044/lc2/view/RegistersPad.java
@@ -1,4 +1,4 @@
-package com.github.h2002044.lc2;
+package com.github.h2002044.lc2.view;
 
 import javax.swing.*;
 import java.awt.*;
